iop/requirements.txt
python-socketio[client]
eventlet
cryptography
fastecdsa
urllib3